Move CheckAPU1SFXPriority to remove a JMP opcode
Now that we don't need to rely on branching distances, the destination has been
de-facto split into two snippets. The one that uses hardcoded 1DFA SFX can end
up directly at SFXTerminateCh, which means that no branching opcode is needed at
all. A little farther down, when a sequence is used for 1DFA SFX, it branches to
ProcessSFXInput, saving a byte there.

A small fix was also done where ProcessAPU1SFX was accidentally called twice
when hardcoded 1DFA SFX was being used, since it was saved in the stack when a
jump table was used.