Move ProcessAPU1SFX to remove a JMP and RET opcode
This technically only applies to hardcoded 1DFA SFX, but since a ret opcode was
relocated, it also affects the non-hardcoded 1DFA SFX slightly, but it doesn't
save a byte there. Enough savings have occurred that a JMP opcode can be removed
and we can go directly to L_0B08 instead. In addition, L_0AB0 is close enough
that we can remove a RET opcode from earlier, since L_0AB0 also points directly
to a ret opcode. So for the hardcoded 1DFA SFX, we save four bytes this way.