Turn an EOR operation into a NOT1 opcode, removing SubC_03
SubC_03 is no longer needed. Instead, we can use a NOT1 opcode, which toggles
the bits directly without the need for an accumulator, and also does not need a
SETP/CLRP pair of opcodes. The other ones that use the TCLR/TSET opcodes are
still valid because of the SETP/CLRP requirement otherwise to use the SET1 and
CLR1 opcodes (and an additional RET opcode).