Merge various nearby patches together affecting the same area
These patches have been merged together as a direct result of either conflicting
with each other by writing to the same memory location or simply being next to
each other. In addition, a change was reverted: the welcome song is stored to
$0DDA instead of $1DFB because the accumulator is overwritten by the level
music, along with a check being missing.

The Bowser fight music skip was never functional because it was overwritten by
one of these patches. I have decided to simply cut it out for the time being.

This commit closes #347.