Skip merging notes and ties on past Addmusics
Rests are allowed to be merged together, but notes and ties are not. This is
because of a conflict with quantization: quantization acts relative to the
length of the last VCMD executed, and ties are separated from normal notes in
this specific situation. Without doing this, notes may key off sooner than
expected compared to the way they were originally created. Rests can still be
merged together because it is already keyed off to begin with.

This commit closes #350.