Use a separate subroutine for the $FD SFX command
We can directly fetch the bytes and send them to the stack to jump immediately
upon execution of a RET opcode instead of using scratch RAM and recycling a JMP
opcode. The number of bytes saved is actually zero as a result of these changes
causing a net offset, partially because of adding a CALL opcode and also because
I kept the X register set to zero for consistency's sake at this time, but a JMP
opcode to .processSFXPitch can instead be removed and the conditional branch
changed, saving three bytes, since it is now barely in range.