Add some commentary for ForceSFXEcho
Indicate what the carry is expected to be in this particular case. It should be
cleared since it came from the jump array. The Yoshi Drum 1DFA commands handle
things differently because of the state of the carry depending on whether or not
the jump array was used or not.