Recycle the accumulator + move SetFLGFromNCKValue after data transfer
The accumulator can be recycled into X as well as not be reloaded for some other
storages that need zeroing out. In addition, SetFLGFromNCKValue was moved to
avoid interfering with the accumulator. These changes end up saving three bytes,
with !PSwitchIsSFX also saving a byte courtesy of using the accumulator there.