Use a word operation for the write to the ESA DSP register
This is clearer to a user when CLR1 and SET1 opcodes do not need to be used to
change the DSP register ID to save memory. There is no save in filesize, but
this one can handle the upcoming changes better since a constant is being
written.

This commit mentions #357.