Add a couple of NOPs into the waiting iteration
The iteration counter that is being used is actually somewhat akin to the number
of bytes written. The echo buffer writes four bytes every 32 cycles, which in
turn means I need to wait eight cycles per iteration. Seven of those are
accounted for via the DBNZ opcode except for the last iteration, which has two
cycles less, so two NOPs are added on so that each iteration is eight cycles or
greater. This will help ensure that the echo buffer is properly captured.

This commit mentions #357.