Update Appveyor configuration for updated uploading methodology
The website that the builds were being uploaded to updated their security
methodology that caused the original method to no longer work. This is the new
updated version that I collaborated with Atari 2.0 (the person hosting the
website containing these auto-compiled builds) so that I can continue to upload
updated builds.