Test that the bit being tested is actually used by hardcoded 1DFA SFX
Somehow triggering two SFX instances in a row while a 1DFA SFX was running was
causing the jump SFX to stop its pitch bend entirely. The cause is that the bit
was not tested to make sure the channel in question was actually used by 1DFA
SFX, so that is now added on.

This commit closes #388.