Filter out all duplicate samples from compiled songs
This commit replicates something that AddmusicK 1.0.8 did with regards to its
sample filtering by simply marking duplicates as empty samples instead in the
song. Notably, this is being done regardless of the important markers.

This commit mentions #396. It does not close it because a longer term fix for
this kind of situation would instead use duplicate pointers, of which the empty
BRR sample does not properly do with its loop point specifically. This is
because at some point down the road the user may want to take advantage of
duplicate pointers allocated to multiple SRCN IDs for whatever reason. This is
also because this currently only affects the C++ side: the full fix would also
get the SNES side involved.