Use a BPL/BMI opcode switch for runningArpGate with MOV1 and NOT1 opcodes
We can save several bytes by taking advantage of the fact that the gate directly
precedes a BPL opcode, meaning it will always branch if set to a BMI opcode.
Most of the savings are small, but one of them is a big one by using a NOT1
opcode, meaning the accumulator doesn't even need to be saved in the stack.