Implement duplicate sample pointers on the C++ side
All SPC files will now compile with sample pointers duplicated instead of
duplicated sample data. Thus, we no longer need to specify an empty sample slot
for a duplicate sample.

This commit mentions #396. It does not close it because the SNES side still
needs to be accounted for, and until that's implemented, compilation will
temporarily be inconsistent.