Recycle a call for writing to KOFF DSP register in .silent
By taking advantage of the accumulator and recycling a call to KeyOffVoices
(though we skip clearing !PlayingVoices), we save two bytes.

This one (and a missed optimization) was also ported to the old unpause silent
SFX ASM that was kept for historical reasons and also to act as an example.