Reduce SFX lag by incorporating the same technique as the music
This is a known vanilla problem with the keyhole song, and like fine tune
interference with the SFX, it eventually had to go. In this case, however, the
reason was that the lag reduction exposed a different problem: namely, that if
the CPU load was heavy enough on the music side, the SFX would also lag. Also as
a side effect of this, it would also affect the polling rate of $1DF9, $1DFA and
$1DFC commands, which is bad for sound driver responsiveness, especially as the
acknowledgement part was taken out.

This commit closes #427.