Disable relative patch filepath warnings when using asar 1.9
This warning is being disabled for asar 1.9 and later because turning relative
paths into absolute paths... might be more complicated than expected. That, and
AddmusicK has a de-facto hardcoded structure, and there may be other relative
paths that the user might use that the program wouldn't catch.

Note that in order to do this, I had to add a command line parameter to the
non-DLL version, and the DLL version was modified to use asar_patch_ex because
there is no way to disable warnings through a standard asar_patch call. On the
plus side, doing this for the DLL version opens it to extra configuration
possibilities down the road.

This commit closes #422.