Discard the hex output entirely and remove 0x marker for stat file tick counts
As much as I would like to have this output hex, the main concern I have is
either integer overflows in the long run if I choose to convert it to an int or
some other shenanigans when trying to handle the conversion of a double for hex
output. So instead, we'll just simply keep the decimal output, and instead get
rid of the hex call and markers to prevent user confusion.

This commit closes #321.