Swap BCS/BEQ opcodes when checking the starman timer for restoration
These two opcodes were checked in the wrong order, causing the BEQ case to never
fire. As for why... the carry is set when greater than or equal to the previous
value.

This code mentions #435. As for why it does not close it, it is because there is
a code snippet at $00E2E5 that may fire at the same time. It also has a snippet
that should be preserved involving a frame counter of some kind in order to
properly preserve the sparkling effect fading out.