Introduce a new song group ID to track the most recent local song loaded
There is a situation where the level music can change during a Starman or
P-Switch effect, and then the music can fail to switch over to the new local
song upon the effect expiring. Since song groups are coming out in the future
and this is one of the cases where it is critical that the correct song group is
loaded, the variable is named to reflect its upcoming purpose.

This commit closes #444.