Initialize ALL DSP registers on boot instead of just a pre-defined list
This will prevent any uninitialized DSP registers from causing shenanigans down
the road. Additionally, this actually saves a few bytes. All of the DSP register
values are covered in their original values except for the DIR DSP register: the
reason why DIR isn't covered is because it is automatically set upon loading a
piece of music, and no samples are currently pre-loaded either.

To avoid a possible accidental out-of-bounds case that occurred in vanilla SMW,
one additional entry has been added to the PanValues table so that $14 is
handled like it was previously. All other out-of-bounds cases will not be
supported as the build changes over time.

This commit closes #453.