Handle SFX termination in the middle of a pitch bend
If the voice has not been keyed off for whatever reason when the zero VCMD is
encountered to end the SFX instance, typically in the middle of a pitch bend,
then all pitch bends are stopped, then SFXTerminateCh is called to effectively
retrigger the command. SFXTerminateCh was actually not used at all if
!useSFXSequenceFor1DFASFX was !true because all of the calls and ways to access
it were filtered out. This also gives !PlayingVoices a purpose other than being
checked during vibrato processing.

This commit closes #460.