Allow $DD VCMD to be called outside of readahead
Took a page out of standard N-SPC and implemented a way to call $DD without
crashing things via zero pointer. I also managed to rig it so that it also
costs no bytes via removing a call to CalcPortamentoDelta and instead go
directly to the routine with no branching required.

We may need to check and make sure this doesn't result in SFX interference down
the road, though.

This commit closes #104.