Deduplicate sample group pointers for songs
Recycling sample group pointers will help reduce the amount of space consumed in
the ROM. It'll also help streamline things in the long run if we were to do lag
reduction on loading new songs, as the simplest form of this involves simply
keeping the sample directory pointer consistent and checking for a duplicate
sample group pointer... although even this would require a data version bump.

This commit closes #424.